+++ configure.py
@@ -686,6 +686,7 @@ scylla_core = (['message/messaging_service.cc',
                 'mutation/frozen_mutation.cc',
                 'mutation/mutation.cc',
                 'mutation/mutation_fragment.cc',
+                'mutation/mutation_fragment_stream_validator.cc',
                 'mutation/mutation_partition.cc',
                 'mutation/mutation_partition_v2.cc',
                 'mutation/mutation_partition_view.cc',
